// apps/front/src/api/model/mediaWithStreamsStreamsItem.ts
/**
 * Generated by otqs v1.0.0 🍺
 * Do not edit manually.
 * @beast/backend
 * OpenAPI spec version: 0.0.1
 */
import type { StreamType } from "./streamType";

export type MediaWithStreamsStreamsItem = {
  id: string;
  mediaId: string | null;
  name: string;
  streamIndex: number;
  type: StreamType;
};